% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cols}
\alias{check_cols}
\title{Check that required columns exist in a dataframe}
\usage{
check_cols(data, cols)
}
\arguments{
\item{data}{A dataframe.}

\item{cols}{Character vector of required column names.}
}
\value{
Invisible TRUE if all columns present, otherwise aborts.
}
\description{
Check that required columns exist in a dataframe
}
\keyword{internal}
