% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{epi_curve}
\alias{epi_curve}
\title{Plot epidemic curve}
\usage{
epi_curve(data, disease = NULL, period = "weekly", fill_color = "#2166ac")
}
\arguments{
\item{data}{A tibble from \code{\link[=epi_clean]{epi_clean()}}.}

\item{disease}{Character. Name of the disease to plot. If NULL, all diseases plotted.}

\item{period}{Character. One of \code{"weekly"} or \code{"monthly"}.}

\item{fill_color}{Character. Bar fill color. Default: \code{"#2166ac"}.}
}
\value{
A ggplot2 object.
}
\description{
Generates a ggplot2 epidemic curve (epi curve) from clean surveillance data.
}
