% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date_range}
\alias{format_date_range}
\title{Format a date range as a readable string}
\usage{
format_date_range(dates)
}
\arguments{
\item{dates}{Date vector.}
}
\value{
Character string like "2024-01-01 to 2024-03-31".
}
\description{
Format a date range as a readable string
}
\keyword{internal}
